fix: Linux build fixes for Qt6
- Fix QString::arg() const char* calls with QString::fromLatin1()
- Fix QImage::flipped() -> mirrored() for Qt6
- Fix Server.cpp sign conversion in protobuf set_video_data()
- Update ScreenShareViewer for Qt6 OpenGL functions

+++ src/murmur/Server.cpp
@@ -1194,7 +1194,7 @@ void Server::processVideoMsg(ServerUser *u, const Mumble::Protocol::VideoData &v
 	videoMsg.set_is_keyframe(videoData.isKeyFrame);
 
 	videoMsg.set_video_data(reinterpret_cast< const char * >(videoData.payload.data()),
-							static_cast< int >(videoData.payload.size()));
+							static_cast< int >(static_cast< long >(videoData.payload.size())));
 	const std::size_t size = videoMsg.ByteSizeLong();
 
 	if (size > static_cast< std::size_t >(std::numeric_limits< int >::max())) {
